config run interface
