+++ main.go
@@ -30,9 +30,9 @@ import (
 	"time"
 
 	"github.com/GoogleCloudPlatform/guest-logging-go/logger"
+	"github.com/GoogleCloudPlatform/osconfig/agentconfig"
 	"github.com/GoogleCloudPlatform/osconfig/agentendpoint"
 	"github.com/GoogleCloudPlatform/osconfig/clog"
-	"github.com/GoogleCloudPlatform/osconfig/config"
 	"github.com/GoogleCloudPlatform/osconfig/policies"
 	"github.com/GoogleCloudPlatform/osconfig/tasker"
 	"github.com/tarm/serial"
@@ -52,9 +52,9 @@ func init() {
 		version = "manual-" + time.Now().Format(time.RFC3339)
 	}
 	// We do this here so the -X value doesn't need the full path.
-	config.SetVersion(version)
+	agentconfig.SetVersion(version)
 
-	os.MkdirAll(filepath.Dir(config.RestartFile()), 0755)
+	os.MkdirAll(filepath.Dir(agentconfig.RestartFile()), 0755)
 }
 
 type serialPort struct {
@@ -77,7 +77,7 @@ var deferredFuncs []func()
 func run(ctx context.Context) {
 	// Setup logging.
 	opts := logger.LogOpts{LoggerName: "OSConfigAgent"}
-	if config.Stdout() {
+	if agentconfig.Stdout() {
 		opts.Writers = []io.Writer{os.Stdout}
 	}
 	if runtime.GOOS == "windows" {
@@ -85,25 +85,25 @@ func run(ctx context.Context) {
 	}
 
 	// If this call to WatchConfig fails (like a metadata error) we can't continue.
-	if err := config.WatchConfig(ctx); err != nil {
+	if err := agentconfig.WatchConfig(ctx); err != nil {
 		logger.Init(ctx, opts)
 		logger.Fatalf(err.Error())
 	}
-	opts.Debug = config.Debug()
-	opts.ProjectName = config.ProjectID()
+	opts.Debug = agentconfig.Debug()
+	opts.ProjectName = agentconfig.ProjectID()
 
 	if err := logger.Init(ctx, opts); err != nil {
 		fmt.Printf("Error initializing logger: %v", err)
 		os.Exit(1)
 	}
-	ctx = clog.WithLabels(ctx, map[string]string{"instance_name": config.Name()})
+	ctx = clog.WithLabels(ctx, map[string]string{"instance_name": agentconfig.Name()})
 
 	// Remove any existing restart file.
-	if err := os.Remove(config.RestartFile()); err != nil && !os.IsNotExist(err) {
+	if err := os.Remove(agentconfig.RestartFile()); err != nil && !os.IsNotExist(err) {
 		clog.Errorf(ctx, "Error removing restart signal file: %v", err)
 	}
 
-	deferredFuncs = append(deferredFuncs, logger.Close, func() { clog.Infof(ctx, "OSConfig Agent (version %s) shutting down.", config.Version()) })
+	deferredFuncs = append(deferredFuncs, logger.Close, func() { clog.Infof(ctx, "OSConfig Agent (version %s) shutting down.", agentconfig.Version()) })
 
 	obtainLock()
 
@@ -115,12 +115,12 @@ func run(ctx context.Context) {
 		}
 	}()
 
-	clog.Infof(ctx, "OSConfig Agent (version %s) started.", config.Version())
+	clog.Infof(ctx, "OSConfig Agent (version %s) started.", agentconfig.Version())
 
 	// Call RegisterAgent on start then at least once every day.
 	go func() {
 		for {
-			if config.TaskNotificationEnabled() || config.GuestPoliciesEnabled() {
+			if agentconfig.TaskNotificationEnabled() || agentconfig.GuestPoliciesEnabled() {
 				if client, err := agentendpoint.NewClient(ctx); err != nil {
 					logger.Errorf(err.Error())
 				} else if err := client.RegisterAgent(ctx); err != nil {
@@ -166,7 +166,7 @@ func runTaskLoop(ctx context.Context, c chan struct{}) {
 	var taskNotificationClient *agentendpoint.Client
 	var err error
 	for {
-		if config.TaskNotificationEnabled() && (taskNotificationClient == nil || taskNotificationClient.Closed()) {
+		if agentconfig.TaskNotificationEnabled() && (taskNotificationClient == nil || taskNotificationClient.Closed()) {
 			// Start WaitForTaskNotification if we need to.
 			taskNotificationClient, err = agentendpoint.NewClient(ctx)
 			if err != nil {
@@ -174,7 +174,7 @@ func runTaskLoop(ctx context.Context, c chan struct{}) {
 			} else {
 				taskNotificationClient.WaitForTaskNotification(ctx)
 			}
-		} else if !config.TaskNotificationEnabled() && taskNotificationClient != nil && !taskNotificationClient.Closed() {
+		} else if !agentconfig.TaskNotificationEnabled() && taskNotificationClient != nil && !taskNotificationClient.Closed() {
 			// Cancel WaitForTaskNotification if we need to, this will block if there is
 			// an existing current task running.
 			if err := taskNotificationClient.Close(); err != nil {
@@ -190,7 +190,7 @@ func runTaskLoop(ctx context.Context, c chan struct{}) {
 			fmt.Println("default")
 		}
 
-		if err := config.WatchConfig(ctx); err != nil {
+		if err := agentconfig.WatchConfig(ctx); err != nil {
 			clog.Errorf(ctx, err.Error())
 		}
 		select {
@@ -210,9 +210,9 @@ func runServiceLoop(ctx context.Context) {
 	<-c
 
 	// Runs functions that need to run on a set interval.
-	ticker := time.NewTicker(config.SvcPollInterval())
+	ticker := time.NewTicker(agentconfig.SvcPollInterval())
 	for {
-		if _, err := os.Stat(config.RestartFile()); err == nil {
+		if _, err := os.Stat(agentconfig.RestartFile()); err == nil {
 			clog.Infof(ctx, "Restart required marker file exists, beginning agent shutdown, waiting for tasks to complete.")
 			tasker.Close()
 			clog.Infof(ctx, "All tasks completed, stopping agent.")
@@ -222,11 +222,11 @@ func runServiceLoop(ctx context.Context) {
 			os.Exit(2)
 		}
 
-		if config.GuestPoliciesEnabled() {
+		if agentconfig.GuestPoliciesEnabled() {
 			policies.Run(ctx)
 		}
 
-		if config.OSInventoryEnabled() {
+		if agentconfig.OSInventoryEnabled() {
 			// This should always run after ospackage.SetConfig.
 			tasker.Enqueue(ctx, "Report OSInventory", func() {
 				client, err := agentendpoint.NewClient(ctx)
@@ -253,7 +253,7 @@ func runServiceLoop(ctx context.Context) {
 func main() {
 	flag.Parse()
 	ctx, cncl := context.WithCancel(context.Background())
-	ctx = clog.WithLabels(ctx, map[string]string{"agent_version": config.Version()})
+	ctx = clog.WithLabels(ctx, map[string]string{"agent_version": agentconfig.Version()})
 	c := make(chan os.Signal, 1)
 	signal.Notify(c, syscall.SIGTERM, syscall.SIGINT)
 	go func() {
